Rename `types::internal` to `types::private` and refer to private types inline as `private::Foo`.
