+++ src/types/binary.rs
@@ -8,7 +8,7 @@ use crate::context::{Context, Lock};
 use crate::handle::Handle;
 use crate::handle::Managed;
 use crate::result::JsResult;
-use crate::types::internal::ValueInternal;
+use crate::types::private::ValueInternal;
 use crate::types::{build, Object, Value};
 use neon_runtime;
 use neon_runtime::raw;
